fix: gfx1151 Strix Halo training path end-to-end
Verified the full bootstrap → smoke-test path on AMD Ryzen AI Max+ 395
(gfx1151) with TheRock nightly torch 2.11.0+rocm7.13. End-to-end SFT
smoke went from broken → 73s/10steps; AOTriton flash backend kicks in
(20x speedup over math fallback on bf16 (2,16,1024,64) attention).

Bootstrap:
- bootstrap.sh: drop torch <2.7 cap (TheRock ships 2.7.0a0+ ... 2.11+)
- bootstrap.sh: replace `uv sync` with `uv pip install -r requirements.txt`
  (uv sync resolves all pyproject extras together; the unsloth extra
  pins old transformers/trl, conflicting with new floors). Auto-skip
  bitsandbytes on ROCm/MPS/CPU (PyPI wheels are CUDA-only).
- bootstrap.sh: write TORCH_ROCM_AOTRITON_ENABLE_EXPERIMENTAL=1 into
  .venv/bin/activate so SDPA picks the AOTriton flash kernel that
  ships inside the TheRock torch wheel (libaotriton_v2.so.0.11.2+).

Trainers:
- train_sft.py: TRL 1.3 renamed SFTConfig.max_seq_length → max_length.
- train_sft.py / train_dpo.py / train_grpo.py: transformers 5.x deprecated
  from_pretrained(torch_dtype=) → dtype=.
- train_sft.py: ROCm override now also sets the AOTriton env var
  defensively and disables packing (TRL warns packing only safe with
  FA2-varlen, which is unavailable on AOTriton sdpa — risk of silent
  cross-sample label contamination).

Configs:
- sft_default.yaml: SmolLM2-360M → SmolLM2-360M-Instruct (base has no
  chat_template, which TRL 1.3 now requires for conversational data).
- sft_default.yaml: warmup_ratio → warmup_steps (transformers 5.2
  removes warmup_ratio).

Tooling:
- Makefile doctor: heredoc was broken under make (each tab-indented
  line ran in its own shell). Moved to scripts/doctor.py.
- doctor: surface TORCH_ROCM_AOTRITON_ENABLE_EXPERIMENTAL value so the
  user can confirm the gate is open.

Docs (verified against ROCm/ROCm#5404, #6034, AOTriton 0.11.2b release,
ROCm/flash-attention README, AuleTechnologies/Aule-Attention):
- rocm-strix-halo SKILL.md: rewrote the flash-attention section. The
  flash_attention_2 attn_implementation in HF Transformers imports the
  upstream flash_attn package, which is CUDA-only; the ROCm fork
  doesn't list gfx1151 and lacks RDNA3 backward kernels. The working
  path is sdpa + AOTriton via the env-var gate. aule-attention's
  Triton backend lists MI300/MI250/MI210/MI100; gfx1151 not on its
  supported list either.
- rocm-strix-halo SKILL.md: documented the 5 known bf16 numerical
  bugs on gfx1151 from ROCm/ROCm#6034 (NaN at small batch / head_dim=32
  / depth>=12 / Adam beta2<0.97 / drift ~step 1000) so future
  recipes plan around them.
- AGENTS.md: platform table row for Strix Halo updated to match.

+++ scripts/train_grpo.py
@@ -44,7 +44,7 @@ def main(cfg: dict[str, Any], args: argparse.Namespace) -> None:
 
     dtype = getattr(torch, model_cfg.get("torch_dtype", "bfloat16"))
     model = AutoModelForCausalLM.from_pretrained(
-        model_cfg["name_or_path"], torch_dtype=dtype,
+        model_cfg["name_or_path"], dtype=dtype,
         attn_implementation=model_cfg.get("attn_implementation", "sdpa"),
     )
